GUI-2539: Move Dimension choices class to views.alarms module
